Feature/sprint 5

+++ light-life/src/Pages/Customer/Reverse/ReserveList.js
@@ -1,9 +1,11 @@
 import React, { useEffect, useState } from "react";
 import firebase from "firebase/app";
 import "firebase/firestore";
+import Swal from "sweetalert2";
 import DietitianData from "../FindDietitians/DietitianData.js";
 import style from "../../../style/reserveList.module.scss";
-
+import image from "../../../style/image.module.scss";
+import nothing from "../../../images/nothing.svg";
 function ReserveList({ reserve, setReserve }) {
   const [index, setIndex] = useState();
   const [dietitians, setDietitians] = useState([]);
@@ -47,18 +49,37 @@ function ReserveList({ reserve, setReserve }) {
   };
   const removeReserveHandler = (e) => {
     const docID = reserve[+e.target.id].reserveID;
-    firebase
-      .firestore()
-      .collection("reserve")
-      .doc(docID)
-      .delete()
-      .then(() => {
-        alert("delete");
-        setReserve([...reserve.filter((r, index) => index !== +e.target.id)]);
-      })
-      .catch((error) => {
-        console.log("Error:", error);
-      });
+    Swal.fire({
+      text: "確定取消預約嗎?",
+      icon: "warning",
+      showCancelButton: true,
+      cancelButtonText: "取消",
+      confirmButtonText: "確定",
+      confirmButtonColor: "#1e4d4e",
+    }).then((res) => {
+      if (res.isConfirmed) {
+        firebase
+          .firestore()
+          .collection("reserve")
+          .doc(docID)
+          .delete()
+          .then(() => {
+            Swal.fire({
+              text: "取消成功",
+              icon: "success",
+              confirmButtonText: "確定",
+              confirmButtonColor: "#1e4d4e",
+            }).then(() => {
+              setReserve([
+                ...reserve.filter((r, index) => index !== +e.target.id),
+              ]);
+            });
+          })
+          .catch((error) => {
+            console.log("Error:", error);
+          });
+      }
+    });
   };
   return (
     <div className={style["reserve-list"]}>
@@ -113,7 +134,9 @@ function ReserveList({ reserve, setReserve }) {
             </div>
           </>
         ) : (
-          <h4>目前沒有預約喔</h4>
+          <div className={image.nothing}>
+            <img src={nothing} />
+          </div>
         )}
       </div>
       <div className={style.checked}>
@@ -170,7 +193,9 @@ function ReserveList({ reserve, setReserve }) {
               )
             )
           ) : (
-            <div>尚未有回覆</div>
+            <div className={image.nothing}>
+              <img src={nothing} />
+            </div>
           )}
         </div>
       </div>
